% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.poll_map}
\alias{plot.poll_map}
\title{Plot a poll_map}
\usage{
\method{plot}{poll_map}(x, ...)
}
\arguments{
\item{x}{A \code{poll_map} object.}

\item{...}{Passed to \code{\link[terra]{plot}}.}
}
\value{
Invisibly \code{NULL}; called for its plotting side effect.
}
\description{
Plot a poll_map
}
\examples{
library(terra)
lulc <- rast(nrows = 20, ncols = 20, xmin = 0, xmax = 2000, ymin = 0, ymax = 2000)
values(lulc) <- 1L; lulc[, 1:3] <- 2L
lc <- poll_landcover(data.frame(lucode = c(1L, 2L),
                                nesting = c(0.1, 0.9), floral = c(0.6, 0.2)))
m <- poll_lonsdorf(lulc, lc, poll_guild(400))
plot(m)
}
