% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.R
\name{poll_demand}
\alias{poll_demand}
\title{Map pollination demand}
\usage{
poll_demand(crop, dependence)
}
\arguments{
\item{crop}{A crop-type \code{SpatRaster} of integer codes.}

\item{dependence}{A \code{\link[=poll_crop_dependence]{poll_crop_dependence()}} table.}
}
\value{
A \code{poll_map} with a single \code{demand} layer in [0, 1].
}
\description{
Converts a crop-type raster into a relative pollination-demand surface by
assigning each crop its pollinator-dependence ratio (Schulp et al. 2014;
Picanco et al. 2017). Non-crop or unmatched cells become 0.
}
\references{
Schulp, C.J.E., Lautenbach, S. & Verburg, P.H. (2014)
\emph{Ecological Indicators}, 36, 131-141.
}
\seealso{
\code{\link[=poll_crop_dependence]{poll_crop_dependence()}}, \code{\link[=poll_balance]{poll_balance()}}
}
