% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{poll_raster}
\alias{poll_raster}
\title{Extract the raster from a poll_map}
\usage{
poll_raster(x)
}
\arguments{
\item{x}{A \code{poll_map} object.}
}
\value{
The underlying \code{terra} \code{SpatRaster}.
}
\description{
Extract the raster from a poll_map
}
\examples{
lulc <- terra::rast(nrows = 10, ncols = 10,
                    xmin = 0, xmax = 1000, ymin = 0, ymax = 1000)
terra::values(lulc) <- 1L
lulc[, 1:2] <- 2L
lc <- poll_landcover(data.frame(lucode = c(1L, 2L),
                                nesting = c(0, 1), floral = c(0.8, 0.1)))
m <- poll_lonsdorf(lulc, lc, poll_guild(alpha = 200))
poll_raster(m)
library(terra)
lulc <- rast(nrows = 20, ncols = 20, xmin = 0, xmax = 2000, ymin = 0, ymax = 2000)
values(lulc) <- 1L; lulc[, 1:3] <- 2L
lc <- poll_landcover(data.frame(lucode = c(1L, 2L),
                                nesting = c(0.1, 0.9), floral = c(0.6, 0.2)))
m <- poll_lonsdorf(lulc, lc, poll_guild(400))
poll_raster(m)
}
