% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.poll_map}
\alias{print.poll_map}
\title{Print a poll_map}
\usage{
\method{print}{poll_map}(x, ...)
}
\arguments{
\item{x}{A \code{poll_map} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a poll_map
}
