% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.poll_map}
\alias{plot.poll_map}
\title{Plot a poll_map}
\usage{
\method{plot}{poll_map}(x, ...)
}
\arguments{
\item{x}{A \code{poll_map} object.}

\item{...}{Passed to \code{\link[terra]{plot}}.}
}
\value{
Invisibly \code{NULL}; called for its plotting side effect.
}
\description{
Plot a poll_map
}
