% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{poll_guild}
\alias{poll_guild}
\title{Define a pollinator guild}
\usage{
poll_guild(alpha, name = "pollinator")
}
\arguments{
\item{alpha}{Mean foraging distance in the same linear map units as the
land-cover raster (typically metres). A single positive number.}

\item{name}{Optional label for the guild.}
}
\value{
An object of class \code{poll_guild}.
}
\description{
Creates a lightweight description of a pollinator species or guild. In this
first release a guild is characterised solely by its mean foraging distance
\code{alpha}; seasonal activity, nesting-substrate preferences and relative
abundance (as used by the full InVEST/Kennedy formulation) will be added to
this object in a later version.
}
\examples{
poll_guild(alpha = 500, name = "Apis")
}
\references{
Greenleaf, S.S., Williams, N.M., Winfree, R. & Kremen, C. (2007) Bee foraging
ranges and their relationship to body size. \emph{Oecologia}, 153, 589-596.
\doi{10.1007/s00442-007-0752-9}
}
\seealso{
\code{\link[=poll_landcover]{poll_landcover()}}, \code{\link[=poll_lonsdorf]{poll_lonsdorf()}}
}
